% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphids.R
\docType{data}
\name{aphids}
\alias{aphids}
\title{Aphid data}
\format{
A list containing whatevs.
}
\usage{
data(aphids)
}
\description{
Insert details about the study
}
\examples{
data(aphids)
}
\references{
insert ref
}
\keyword{datasets}
