% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphidsBACI.R
\docType{data}
\name{aphidsBACI}
\alias{aphidsBACI}
\title{Aphid data as a BACI design}
\format{
A dataframe containing:\describe{
\item{Plot}{The plot ID, a factor with eight levels}
\item{Treatment}{A factor indicating whether birds were excluded, \code{excluded} or \code{present}.}
\item{Time}{The data of sampling (June 18th or July 15th, 3 and 30 days after netting, respectively)}
\item{counts}{Aphid abundance (counts)}
\item{logcount}{log(y+1)-transformed aphid abundance}
}
}
\usage{
data(aphidsBACI)
}
\description{

}
\examples{
data(aphidsBACI)
plot(logcount~interaction(Treatment,Time),data=aphidsBACI,col=c("lightgreen","pink"))
}
\references{
Grass \emph{et al.} (2017) Insectivorous birds disrupt biological control of cereal aphids.
Ecology, \bold{98} 1583-90.
}
\keyword{datasets}
