% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamenvelope.R
\name{gamenvelope}
\alias{gamenvelope}
\title{Add a scatterplot with a GAM smoother and confidence bands}
\usage{
gamenvelope(
  x,
  y,
  conf.level = 0.95,
  col = par("col"),
  envelope.col = c("blue", rgb(0, 0, 1, 0.25)),
  ...
)
}
\arguments{
\item{x}{the \code{x} coordinates for the plot.}

\item{y}{the \code{y} coordinates for the plot.}

\item{conf.level}{the confidence level to use in constructing the point-wise 
confidence bands.}

\item{col}{the colors for points. Multiple colors can be specified so that each point
can be given its own color. If there are fewer colors than points they are recycled
in the standard fashion.}

\item{envelope.col}{a character vector of length two whose first and second values 
respectively specify the color of the smoother and the color of the confidence band.}

\item{...}{other graphical parameters passed through to \code{\link{points}}}
}
\value{
A list with the following components:
\item{xSorted}{A vector of \code{x} coordinates sorted from smallest to largest}
\item{prY    }{Predicted values for \code{y} from the GAM smoother, as a function of \code{xSorted}}
\item{prLow  }{Lower confidence band on the GAM smoother}
\item{prHi   }{Upper confidence band on the GAM smoother}
\item{gam.yx }{The fitted \code{gam} object}
}
\description{
A scatterplot with a GAM smoother and confidence bands.
}
\details{
A scatterplot function which adds a GAM smoother and confidence band to the plot.
This is especially useful as a diagnostic tool for residual vs fits plots. The smoother
is fitted using \code{\link[mgcv]{gam}} and point-wise confidence bands use Wald intervals
(adding and subtracting a \code{qnorm(1-(1-conf.level)/2)} standard errors from predictions).
}
\examples{
data(iris)
# fit a mlm:
iris.mlm=lm(cbind(Sepal.Length,Sepal.Width,Petal.Length,Petal.Width)~Species,data=iris)
# construct a residual vs fits plot, with GAM smoother and confidence bands:
plot(iris.mlm,which=1,panel=gamenvelope)

}
\seealso{
\code{\link{lm}}, \code{\link{plot.mlm}}, \code{\link{predict.mlm}}, \code{\link{rstandard}}
}
\author{
David Warton <david.warton@unsw.edu.au>
}
