% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operacion_compleja.R
\name{operacion_compleja}
\alias{operacion_compleja}
\title{A operation Function}
\usage{
operacion_compleja(x)
}
\arguments{
\item{x}{the number that you want to use}
}
\description{
This function allows you to do a simple operation
}
\examples{
operacion_simple()
}
\keyword{operacion}
