% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operacion_simple.R
\name{operacion_simple}
\alias{operacion_simple}
\title{A operation Function}
\usage{
operacion_simple(x)
}
\arguments{
\item{x}{the number that you want to use}
}
\description{
This function allows you to do a little operation
}
\examples{
operacion_simple()
}
\keyword{operacion}
