% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mlm.R
\name{plot.mlm}
\alias{plot.mlm}
\title{Plot Diagnostics for a Multivariate Linear Model Object}
\usage{
\method{plot}{mlm}(
  x,
  which = c(1L:3L, 5L),
  overlay = TRUE,
  caption = list("Residuals vs Fitted", "Normal Q-Q", "Scale-Location",
    "Cook's distance", "Residuals vs Leverage", expression("Cook's dist vs Leverage  " *
    h[ii]/(1 - h[ii]))),
  panel = if (add.smooth == T) gamenvelope else points,
  col = if (overlay) 1:dim(x$model[[1]])[2] else 1,
  sub.caption = NULL,
  main = "",
  ask = prod(par("mfcol")) < length(which) && dev.interactive(),
  ...,
  id.n = 3,
  labels.id = names(residuals(x)),
  cex.id = 0.75,
  qqline = TRUE,
  cook.levels = c(0.5, 1),
  add.smooth = TRUE,
  label.pos = c(4, 2),
  cex.caption = 1,
  cex.oma.main = 1.25
)
}
\arguments{
\item{x}{a \code{mlm} object, typically the result of calling \code{lm} with a matrix response.}

\item{which}{as for \code{\link{plot.lm}}. If a subset of the plots is required, 
specify a subset of the numbers 1:6, see caption below (and the ‘Details’)
for the different kinds.}

\item{overlay}{logical \code{TRUE} (default) means responses are overlaid on a single plot,
after standardizing to ensure they are all on a comparable scale. Otherwise separate plots
are generated for each response variable.}

\item{caption}{captions to appear above the plots, of length 6, corresponding to the 6 potential plots.}

\item{panel}{panel function. Default is \code{\link{gamenvelope}} to plot smoothers with
confidence bands.}

\item{col}{color of points. When responses are overlaid on a single plot, each response
has a different color.}

\item{sub.caption}{common title across figures. If \code{NULL}, as by default, a possibly
abbreviated version of \code{deparse(x$call)} is used.}

\item{main}{title to each plot -- in addition to \code{caption}.}

\item{ask}{logical; if TRUE, the user is asked before each plot, see \code{par(ask=.)}.}

\item{...}{other parameters to be passed through to plotting functions.}

\item{id.n}{number of points to be labelled in each plot, starting with the most extreme.}

\item{labels.id}{vector of labels, from which the labels for extreme points will be chosen.
NULL uses observation numbers.}

\item{cex.id}{magnification of point labels.}

\item{qqline}{logical indicating if a qqline() should be added to the normal Q-Q plot.}

\item{cook.levels}{levels of Cook's distance at which to draw contours.}

\item{add.smooth}{logical indicating if a smoother should be added to most plots; see also panel above.}

\item{label.pos}{positioning of labels, for the left half and right half of the graph respectively, for plots 1-3.}

\item{cex.caption}{controls the size of caption.}

\item{cex.oma.main}{controls the size of the sub.caption only if that is above the figures when there is more than one.}
}
\value{
a graph or set of graphs
}
\description{
A plot function for \code{mlm} objects, constructing diagnostic plots as for \code{lm}.
Iff multivariate linear model assumptions are satisfied, each response satisfies a linear
model as a function of all other response variables as well as other predictors. Hence
we generate residual diagnostics for each of these linear models (and optionally overlay 
them on a single plot).

Borrows heavily from \code{\link{plot.lm}}, with just a few cosmetic changes.
}
\details{
This function borrows heavily from \code{\link{plot.lm}} so see that file for usage details.
The essential difference is that residuals and fitted values are calculated from the
`full conditional` model, that is, as a linear function of all other response variables as 
well as a function of predictors. The full conditionals of a model are diagnostic of the
joint distribution, and in the special case of a multivariate linear model, this means that
the model is correct if each response satisfies a linear model as a function of all other
responses and other predictors. Most of the work is done in \code{\link{predict.mlm}} and
\code{\link{residuals.mlm}} functions.
}
\examples{
data(iris)
# fit a mlm:
iris.mlm=lm(cbind(Sepal.Length,Sepal.Width,Petal.Length,Petal.Width)~Species,data=iris)
# construct residual vs fits and QQ plot
plot(iris.mlm, which=1:2)

}
\seealso{
\code{\link{lm}}, \code{\link{residuals.mlm}}, \code{\link{predict.mlm}}, \code{\link{plot.lm}}
}
\author{
David Warton <david.warton@unsw.edu.au> and Christopher Chung
}
