% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mlm.R
\name{predict.mlm}
\alias{predict.mlm}
\title{Predict method for Multivariate Linear Model Fits}
\usage{
\method{predict}{mlm}(object, conditional = FALSE, standardize = FALSE, ...)
}
\arguments{
\item{object}{a \code{mlm} object, typically the result of calling \code{lm} with a matrix response.}

\item{conditional}{logical defaults to \code{FALSE} for the usual marginal predictions, but
for \code{TRUE} this returns predictions for each respones conditional on the values of
all other responses.}

\item{standardize}{logical defaults to \code{FALSE}, but for \code{TRUE} predictions will be
standardised so they are comparable across responses.}

\item{...}{further arguments passed to \code{\link{predict.lm}}, in particular, \code{newdata}.
However, this function was not written to accept non-default values for \code{se.fit}, 
\code{interval} or \code{terms}.}
}
\value{
A matrix of predicted values.
}
\description{
Predicted values based on a multivariate linear model (\code{mlm}) object.
Returns the usual (marginal) predictions by default, but optionally returns
predictions from a full conditional model (\code{conditional=TRUE}), which
can be used to diagnose the multivariate normality assumption.
}
\details{
A \code{predict} function for \code{mlm} objects, which returns the usual (marginal) 
predicted values by default, but optionally returns predictions from a full conditional 
model (\code{conditional=TRUE}), that is, from a linear model for each response as a function
of all responses as well as predictors. This can be used in plots to diagnose the
multivariate normality assumption.

These can be standardized (\code{standardize=TRUE}) to facilitate overlay plots of multiple
responses, as is the default behaviour in \code{\link{plot.mlm}}.

Currently, standard errors and confidence intervals around predictions are not available.
}
\examples{
data(iris)
# fit a mlm:
iris.mlm=lm(cbind(Sepal.Length,Sepal.Width,Petal.Length,Petal.Width)~Species,data=iris)
# predict each response conditionally on the values of all other responses:
predict(iris.mlm, conditional=TRUE)

}
\seealso{
\code{\link{lm}}, \code{\link{plot.mlm}}, \code{\link{predict.lm}}, \code{\link{residuals.mlm}}
}
\author{
David Warton <david.warton@unsw.edu.au>
}
