% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.mlm.R
\name{residuals.mlm}
\alias{residuals.mlm}
\title{Extract Multivariate Linear Model Residuals}
\usage{
\method{residuals}{mlm}(object, conditional = FALSE, standardize = FALSE, ...)
}
\arguments{
\item{object}{a \code{mlm} object, typically the result of calling \code{lm} with a matrix response.}

\item{conditional}{logical defaults to \code{FALSE} for the usual marginal residuals, but
for \code{TRUE} this returns residuals from a linear model of each response against all
other responses, together with other model predictors.}

\item{standardize}{logical defaults to \code{FALSE}, but for \code{TRUE} residuals will be
studentized using \code{\link{rstandard}} so they are comparable across responses.}

\item{...}{further arguments passed to \code{\link{residuals.lm}} or \code{\link{rstandard}}.}
}
\value{
A matrix of residuals
}
\description{
Residuals for a Multivariate Linear Model (\code{mlm}) object. Returns the usual (marginal) 
residuals by default, but optionally returns residuals from a full conditional model 
(\code{conditional=TRUE}), which can be used to diagnose the multivariate normality 
assumption.
}
\details{
A \code{residuals} function for \code{mlm} objects, which returns the usual (marginal) 
residuals by default, but optionally returns residuals from a full conditional model 
(\code{conditional=TRUE}), that is, a linear model of each response against all responses
as well as predictors, which can be used to diagnose the multivariate normality assumption.
These can be standardized (\code{standardize=TRUE}) to facilitate overlay plots of multiple
responses, as is the default behaviour in \code{\link{plot.mlm}}.
}
\examples{
data(iris)
# fit a mlm:
iris.mlm=lm(cbind(Sepal.Length,Sepal.Width,Petal.Length,Petal.Width)~Species,data=iris)
# construct full conditional residuals:
residuals(iris.mlm, conditional=TRUE)

}
\seealso{
\code{\link{lm}}, \code{\link{plot.mlm}}, \code{\link{predict.mlm}}, \code{\link{rstandard}}
}
\author{
David Warton <david.warton@unsw.edu.au>
}
